# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ODBUS Consumer APIs",
        "description": "L5 Swagger OpenApi description for ODBUS Consumer APIs",
        "version": "0.1"
    },
    "paths": {
...
        "/api/JourneyDetails": {
            "post": {
                "tags": [
                    "Journey details of a costumer"
                ],
                "summary": "Journey Details",
                "description": "Journey Details",
                "operationId": "3d17c3033167f8042e49151b2756642a",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ODM5163863"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "9090909090"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get all Journey details"
                    },
                    "206": {
                        "description": "Validation error: Not a valid pnr or Mobile number"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/PassengerDetails": {
            "post": {
                "tags": [
                    "Passenger details"
                ],
                "summary": "Passenger Details",
                "description": "Passenger Details",
                "operationId": "23595563c399d5db365e5639cfc0e0ba",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ODM5163863"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "9090909090"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get all Passenger details"
                    },
                    "206": {
                        "description": "Validation error: Not a valid pnr or Mobile number"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/BookingDetails": {
            "post": {
                "tags": [
                    "Booking details of a customer"
                ],
                "summary": "Booking Details",
                "description": "Booking Details",
                "operationId": "df359a886d3c0ab31ddfe062aa632b5e",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ODM5163863"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "9090909090"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "get all Booking details of a customer"
                    },
                    "206": {
                        "description": "Validation error: Not a valid pnr or Mobile number"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/EmailSms": {
            "post": {
                "tags": [
                    "Resending Ticket via Email/Sms"
                ],
                "summary": "Resending Ticket via Email/Sms",
                "description": "Resending Ticket via Email/Sms",
                "operationId": "9f16c6a3c6e568a5c7b5987d89555311",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ODM5163863"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "9090909090"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "send email/sms"
                    },
                    "206": {
                        "description": "Validation error: Not a valid pnr or Mobile number"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/cancelTicketInfo": {
            "post": {
                "tags": [
                    "Get detail information on cancellation rules and policies"
                ],
                "summary": "Get detail information on cancellation rules and policies",
                "description": "Get detail information on cancellation rules and policies",
                "operationId": "be66a3872e42dedf728eeae29f225439",
                "parameters": [
                    {
                        "name": "pnr",
                        "in": "query",
                        "description": "pnr number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "ODM5163863"
                        }
                    },
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "mobile number",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "9090909090"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get detail information of ticket cancellation rules and policies"
                    },
                    "206": {
                        "description": "Validation error: Not a valid pnr or Mobile number"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "No record found"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "502": {
                        "description": "Bad gateway"
                    },
                    "503": {
                        "description": "Service unavailable"
                    },
                    "504": {
                        "description": "Gateway timeout"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "name": "Token based",
                "in": "header",
                "bearerFormat": "AUTH0",
                "scheme": "bearer"
            }
        }
    }
}